requirements.txt
langchain
langchain_community
langchain_core
langchain_openai
pydantic
pydantic-settings
langgraph
pypdf
faiss-cpu
fastapi
uvicorn[standard]
python-dotenv
python-multipart